Add a helper to create SA kubeconfigs
We need serviceaccount kubeconfigs for a variety of use cases. Make this
simple to encourage creating distinct kubeconfigs rather than using the
admin kubeconfig everywhere and de-duplicate the existing occurences of
this.
