+++ control-plane-operator/controllers/hostedcontrolplane/pki/kas.go
@@ -5,9 +5,13 @@ import (
 	"net"
 
 	corev1 "k8s.io/api/core/v1"
+	"k8s.io/client-go/tools/clientcmd"
 
+	hyperv1 "github.com/openshift/hypershift/api/v1alpha1"
 	"github.com/openshift/hypershift/control-plane-operator/controllers/hostedcontrolplane/manifests"
 	"github.com/openshift/hypershift/support/config"
+	"github.com/openshift/hypershift/support/util"
+	clientcmdapi "k8s.io/client-go/tools/clientcmd/api"
 )
 
 const (
@@ -66,12 +70,14 @@ func ReconcileKASAdminClientCertSecret(secret, ca *corev1.Secret, ownerRef confi
 	return reconcileSignedCert(secret, ca, ownerRef, "system:admin", []string{"system:masters"}, X509UsageClientServerAuth)
 }
 
-func ReconcileIngressOperatorClientCertSecret(secret, ca *corev1.Secret, ownerRef config.OwnerRef) error {
-	return reconcileSignedCert(secret, ca, ownerRef, "system:serviceaccount:openshift-ingress-operator:ingress-operator", []string{"system:serviceaccounts"}, X509UsageClientServerAuth)
-}
+func ReconcileServiceAccountKubeconfig(secret, ca *corev1.Secret, hcp *hyperv1.HostedControlPlane, serviceAccountNamespace, serviceAccountName string) error {
+	cn := "system:serviceaccount:" + serviceAccountNamespace + ":" + serviceAccountName
+	if err := reconcileSignedCert(secret, ca, config.OwnerRef{}, cn, []string{"system:serviceaccounts"}, X509UsageClientServerAuth); err != nil {
+		return fmt.Errorf("failed to reconcile serviceaccount client cert: %w", err)
+	}
+	svcURL := inClusterKASURL(hcp.Namespace, util.APIPortWithDefault(hcp, config.DefaultAPIServerPort))
 
-func ReconcileAWSPodIdentityWebhookClientCertSecret(secret, ca *corev1.Secret, ownerRef config.OwnerRef) error {
-	return reconcileSignedCert(secret, ca, ownerRef, "system:serviceaccount:openshift-authentication:aws-pod-identity-webhook", []string{"system:serviceaccounts"}, X509UsageClientServerAuth)
+	return ReconcileKubeConfig(secret, secret, ca, svcURL, "", manifests.KubeconfigScopeLocal, config.OwnerRef{})
 }
 
 func nextIP(ip net.IP) net.IP {
@@ -92,4 +98,58 @@ func firstIP(network *net.IPNet) net.IP {
 
 func isNumericIP(s string) bool {
 	return net.ParseIP(s) != nil
+}
+
+func ReconcileKubeConfig(secret, cert, ca *corev1.Secret, url string, key string, scope manifests.KubeconfigScope, ownerRef config.OwnerRef) error {
+	ownerRef.ApplyTo(secret)
+	caBytes := ca.Data[CASignerCertMapKey]
+	crtBytes, keyBytes := cert.Data[corev1.TLSCertKey], cert.Data[corev1.TLSPrivateKeyKey]
+	kubeCfgBytes, err := generateKubeConfig(url, crtBytes, keyBytes, caBytes)
+	if err != nil {
+		return fmt.Errorf("failed to generate kubeconfig: %w", err)
+	}
+	if secret.Data == nil {
+		secret.Data = map[string][]byte{}
+	}
+	if key == "" {
+		key = util.KubeconfigKey
+	}
+	if secret.Labels == nil {
+		secret.Labels = map[string]string{}
+	}
+	secret.Labels[manifests.KubeconfigScopeLabel] = string(scope)
+	secret.Data[key] = kubeCfgBytes
+	return nil
+}
+
+func generateKubeConfig(url string, crtBytes, keyBytes, caBytes []byte) ([]byte, error) {
+	kubeCfg := clientcmdapi.Config{
+		Kind:       "Config",
+		APIVersion: "v1",
+	}
+	kubeCfg.Clusters = map[string]*clientcmdapi.Cluster{
+		"cluster": {
+			Server:                   url,
+			CertificateAuthorityData: caBytes,
+		},
+	}
+	kubeCfg.AuthInfos = map[string]*clientcmdapi.AuthInfo{
+		"admin": {
+			ClientCertificateData: crtBytes,
+			ClientKeyData:         keyBytes,
+		},
+	}
+	kubeCfg.Contexts = map[string]*clientcmdapi.Context{
+		"admin": {
+			Cluster:   "cluster",
+			AuthInfo:  "admin",
+			Namespace: "default",
+		},
+	}
+	kubeCfg.CurrentContext = "admin"
+	return clientcmd.Write(kubeCfg)
+}
+
+func inClusterKASURL(namespace string, apiServerPort int32) string {
+	return fmt.Sprintf("https://%s:%d", manifests.KubeAPIServerServiceName, apiServerPort)
 }
